RBAC in workspace. Completion of Organization and Workspace lifecycles

+++ app/Actions/Workspaces/DeleteWorkspaceAction.php
@@ -13,17 +13,30 @@ class DeleteWorkspaceAction
     ) {
         //
     }
-    public function execute(Workspace $workspace)
+    public function handle(Workspace $workspace)
     {
         if ($workspace->is_default) {
             throw new DomainException('Default workspace cannot be deleted.');
         }
+
+        if ($workspace->teams()->exists()) {
+            throw new DomainException(
+                'Delete all teams before deleting this workspace.'
+            );
+        }
+
+        if ($workspace->projects()->exists()) {
+            throw new DomainException(
+                'Delete all projects before deleting this workspace.'
+            );
+        }
+
         $workspace->delete();
 
         //activity log
         $this->activity->handle(
             event: "{$workspace->name} has been deleted",
-            subject: $workspace->id,
+            subject: $workspace,
             properties: [
                 'workspace_name' => $workspace->name,
             ]
